Feat: Update to v0.4.0 - CU-86c6yv89p
feat: Update dependencies and improve cross-platform file loading

This release introduces support for conditional imports for the `IOFileLoader` to ensure compatibility with web platforms where `dart:io` is unavailable.

Other changes include:
- Bumping `playx_core` to `^1.0.0`.
- Updating `flutter_secure_storage` to `v10.0.0` and migrating macOS/iOS integration to use `flutter_secure_storage_darwin`.
- Bumping SDK constraints to Dart `^3.9.0` and Flutter `^3.35.0`.
- Updating various internal dependencies including `get_it`, `shared_preferences`, and `meta`.
- Adding new localization features: fallback translations for empty resources, `scriptCode` support in `XLocale`, deep translation merging, and device sync reset.
- Updating package version to `0.4.0`.
